+++ internal/layer/benchmark_test.go
@@ -9,16 +9,16 @@ import (
 )
 
 // fillRandom fills a slice with random values.
-func fillRandom(slice []float64) {
+func fillRandom(slice []float32) {
 	for i := range slice {
-		slice[i] = rand.Float64()
+		slice[i] = rand.Float32()
 	}
 }
 
 // BenchmarkDenseForward benchmarks the forward pass of a dense layer.
 func BenchmarkDenseForward(b *testing.B) {
 	layer := NewDense(784, 256, activations.Tanh{})
-	input := make([]float64, 784)
+	input := make([]float32, 784)
 	fillRandom(input)
 
 	b.ResetTimer()
@@ -30,8 +30,8 @@ func BenchmarkDenseForward(b *testing.B) {
 // BenchmarkDenseBackward benchmarks the backward pass of a dense layer.
 func BenchmarkDenseBackward(b *testing.B) {
 	layer := NewDense(784, 256, activations.Tanh{})
-	input := make([]float64, 784)
-	grad := make([]float64, 256)
+	input := make([]float32, 784)
+	grad := make([]float32, 256)
 	fillRandom(input)
 	fillRandom(grad)
 
@@ -47,8 +47,8 @@ func BenchmarkDenseBackward(b *testing.B) {
 // BenchmarkDenseFull benchmarks a complete forward and backward pass.
 func BenchmarkDenseFull(b *testing.B) {
 	layer := NewDense(784, 256, activations.Tanh{})
-	input := make([]float64, 784)
-	grad := make([]float64, 256)
+	input := make([]float32, 784)
+	grad := make([]float32, 256)
 	fillRandom(input)
 	fillRandom(grad)
 
@@ -62,7 +62,7 @@ func BenchmarkDenseFull(b *testing.B) {
 // BenchmarkDenseLarge benchmarks a large dense layer (typical MLP hidden layer).
 func BenchmarkDenseLarge(b *testing.B) {
 	layer := NewDense(1024, 1024, activations.Tanh{})
-	input := make([]float64, 1024)
+	input := make([]float32, 1024)
 	fillRandom(input)
 
 	b.ResetTimer()
@@ -76,7 +76,7 @@ func BenchmarkConv2DForward(b *testing.B) {
 	// Typical CNN configuration: 3 input channels, 64 output channels, 3x3 kernel
 	// Input: 32x32x3 = 3072 elements
 	layer := NewConv2D(3, 64, 3, 1, 1, activations.Tanh{})
-	input := make([]float64, 3*32*32)
+	input := make([]float32, 3*32*32)
 	fillRandom(input)
 
 	b.ResetTimer()
@@ -89,8 +89,8 @@ func BenchmarkConv2DForward(b *testing.B) {
 func BenchmarkConv2DBackward(b *testing.B) {
 	// Typical CNN configuration
 	layer := NewConv2D(3, 64, 3, 1, 1, activations.Tanh{})
-	input := make([]float64, 3*32*32)
-	grad := make([]float64, 64*32*32) // Output: 64 channels, 32x32
+	input := make([]float32, 3*32*32)
+	grad := make([]float32, 64*32*32) // Output: 64 channels, 32x32
 	fillRandom(input)
 	fillRandom(grad)
 
@@ -106,8 +106,8 @@ func BenchmarkConv2DBackward(b *testing.B) {
 // BenchmarkConv2DFull benchmarks a complete forward and backward pass.
 func BenchmarkConv2DFull(b *testing.B) {
 	layer := NewConv2D(3, 64, 3, 1, 1, activations.Tanh{})
-	input := make([]float64, 3*32*32)
-	grad := make([]float64, 64*32*32)
+	input := make([]float32, 3*32*32)
+	grad := make([]float32, 64*32*32)
 	fillRandom(input)
 	fillRandom(grad)
 
@@ -123,7 +123,7 @@ func BenchmarkConv2DLarge(b *testing.B) {
 	// Larger CNN: 64 input channels, 128 output channels, 3x3 kernel
 	// Input: 64x64x64 = 262144 elements
 	layer := NewConv2D(64, 128, 3, 1, 1, activations.Tanh{})
-	input := make([]float64, 64*64*64)
+	input := make([]float32, 64*64*64)
 	fillRandom(input)
 
 	b.ResetTimer()
@@ -135,7 +135,7 @@ func BenchmarkConv2DLarge(b *testing.B) {
 // BenchmarkLSTMForward benchmarks the forward pass of an LSTM layer.
 func BenchmarkLSTMForward(b *testing.B) {
 	layer := NewLSTM(256, 512) // Typical LSTM config
-	input := make([]float64, 256)
+	input := make([]float32, 256)
 	fillRandom(input)
 
 	b.ResetTimer()
@@ -147,8 +147,8 @@ func BenchmarkLSTMForward(b *testing.B) {
 // BenchmarkLSTMBackward benchmarks the backward pass of an LSTM layer.
 func BenchmarkLSTMBackward(b *testing.B) {
 	layer := NewLSTM(256, 512)
-	input := make([]float64, 256)
-	grad := make([]float64, 512)
+	input := make([]float32, 256)
+	grad := make([]float32, 512)
 	fillRandom(input)
 	fillRandom(grad)
 
@@ -164,8 +164,8 @@ func BenchmarkLSTMBackward(b *testing.B) {
 // BenchmarkLSTMFull benchmarks a complete forward and backward pass.
 func BenchmarkLSTMFull(b *testing.B) {
 	layer := NewLSTM(256, 512)
-	input := make([]float64, 256)
-	grad := make([]float64, 512)
+	input := make([]float32, 256)
+	grad := make([]float32, 512)
 	fillRandom(input)
 	fillRandom(grad)
 
@@ -179,8 +179,8 @@ func BenchmarkLSTMFull(b *testing.B) {
 // BenchmarkLSTMSequence benchmarks processing a sequence of 10 timesteps.
 func BenchmarkLSTMSequence(b *testing.B) {
 	layer := NewLSTM(256, 512)
-	input := make([]float64, 256)
-	grad := make([]float64, 512)
+	input := make([]float32, 256)
+	grad := make([]float32, 512)
 	fillRandom(input)
 	fillRandom(grad)
